+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 We follow [Semantic Versions](https://semver.org/).
 
 
+## Version 0.1.1
+
+### Bugfixes
+
+- Changes `puncuation` class name to `punctuation`
+- Changes `className` to `scope`
+- Fixes how `const` is highlighted
+- Fixes URLs from `sobolevn/` to `highlightjs/`
+
+
 ## Version 0.1.0
 
 - Initial release